Making FABridge global variables really global.
When concatenating FABridge with other scripts into one big .js file,
it can happen that the FABridge.js code is no more run in global scope,
and when that happens the functions defined in FABridge.js won't be
global, most importantly those functions that Flash wants to call
that really have to be global.

The standard practice in JavaScript libraries is to add that kind
of functions directly to the window object.  This is what I have
done here.

Additonally renamed instanceFactory() into FABridge__instanceFactory().
And made the accidental gobal variable newType into local variable.